requirements.txt
python-docx==1.1.2
python-pptx==1.0.2
PyPDF2==3.0.1
pytesseract==0.3.13
opencv-python==4.11.0.86
pymupdf==1.24.11
scikit-learn==1.3.2
watchdog==4.0.2
pystray==0.19.5
winrt==1.0.21033.1